refactor: Piece 타입에 비어있는 칸인 Empty 객체 추가
