+++ ChessApp/ChessApp/Error/BoardError.swift
@@ -10,6 +10,7 @@ import Foundation
 enum BoardError: Error {
   case notMovableFile
   case notMovableRank
+  case emptyBoard
   
   var message: String {
     switch self {
@@ -17,6 +18,8 @@ enum BoardError: Error {
       return "움직일 수 있는 File칸이 존재하지 않습니다."
     case .notMovableRank:
       return "움직일 수 있는 Rank칸이 존재하지 않습니다."
+    case .emptyBoard:
+      return "비어있는 칸입니다."
     }
   }
 }
